batt_smbus: correct args to I2C transfer
Also added read_reg method to reduce duplicate code
add retry to I2C init and register reads

+++ src/drivers/drv_batt_smbus.h
@@ -71,4 +71,4 @@
 #define BATT_SMBUS_READ_CELL_VOLTAGE	_BATT_SMBUS_IOC(5)
 
 /* structure to hold individual cell voltages */
-typedef unsigned int batt_smbus_cell_voltage[BATT_SMBUS_MAX_CELLS];
+typedef uint16_t batt_smbus_cell_voltage[BATT_SMBUS_MAX_CELLS];
